requirements.txt
numpy==1.16.5
matplotlib==3.1.1
ipywidgets==7.2.1
bqplot==0.11.6
platypus-opt
netcdf4==1.4.2
numba==0.47.0
plotly==4.4.1
cdsapi
pandas==0.25.1